refactored liquid transformation code (has some flaws)
Completely rewrote the liquid transformation. The algorithm now exclusiv
ely modifies the current node in one iteration. Another change is that M
apNode.param2 now stores a bit that tells other nodes if a flowing liqui
d node flows downwards. This is accomplished by two masks on param2 for
the "flow down" bit and the "liquid level" bits.

This will be the base
of future improvements to the liquid flow. However, in the current state
 flowing liquid does not always disappear when cut off its source. Be aw
are that this branch is work in progress.